% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acc_topk}
\alias{acc_topk}
\title{get top k accuracy}
\usage{
acc_topk(pred_topk, truth)
}
\arguments{
\item{pred_topk}{a matrix of integers; each row indicates top 1, 2, ..., k
causes}

\item{truth}{a vector of integers; its length equals the number of rows in
\code{pred_topk}.}
}
\value{
numeric; between 0 and 1; the higher the more accurate.
}
\description{
get top k accuracy
}
\examples{

xx <- matrix(c(0.3,0.1,0.2,0.4,
               0.4,0.1,0.3,0.2),nrow=2,ncol=4,byrow=TRUE)
pred_top2 <- get_topk_COD(xx,2)
acc_topk(pred_top2,c(4,2))

}
