% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletree_example_data.R
\docType{data}
\name{cause_ids_except_AP}
\alias{cause_ids_except_AP}
\title{example cause ids with NAs}
\format{
vector; the length is 7841
}
\usage{
cause_ids_except_AP
}
\description{
a vector of character strings
}
\keyword{datasets}
