% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletree_example_data.R
\docType{data}
\name{domain_ids}
\alias{domain_ids}
\title{example cause ids}
\format{
vector; the length is 7841
}
\usage{
domain_ids
}
\description{
a vector of character strings
}
\keyword{datasets}
