% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{doubletree_blca}
\alias{doubletree_blca}
\title{simulate data and subject-specific
class indicators from latent class models}
\usage{
doubletree_blca(n, itemprob = 0.5, classprob = 1)
}
\arguments{
\item{n}{sample size}

\item{itemprob}{item probabilities; K by J.}

\item{classprob}{class probabilities; K.}
}
\value{
a list; \code{x}: data; \code{z} a vector of integer indicators
of class membership.
}
\description{
simulate data and subject-specific
class indicators from latent class models
}
\examples{

doubletree_blca(10,matrix(0.5,nrow=3,ncol=20),c(0.98,0.01,0.01))

}
\seealso{
\code{\link[BayesLCA:blca]{BayesLCA::blca()}}
}
