% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletree_example_data.R
\docType{data}
\name{example_data_doubletree}
\alias{example_data_doubletree}
\title{simulated data (K=2,J=20,N=1000)}
\format{
matrix; 1000 by 20
}
\usage{
example_data_doubletree
}
\description{
a matrix of simulated data
\code{1}: yes; \code{0}: no
}
\keyword{datasets}
