% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{F_doubletree}
\alias{F_doubletree}
\title{Calculate the F term; an array}
\usage{
F_doubletree(
  psi,
  g_psi,
  phi,
  g_phi,
  X,
  ind_obs_i,
  rmat,
  E_beta,
  E_beta_sq,
  E_eta,
  E_eta_sq,
  v1_lookup_NA_replaced,
  v2_lookup
)
}
\arguments{
\item{psi, g_psi, phi, g_phi}{local variational parameters}

\item{X}{transformed data: \verb{2Y-1}; with potential missing responses}

\item{ind_obs_i}{List of length N; each element is a vector of integers; variable lengths.}

\item{rmat}{matrix (N by K); row sums are one - the class probabilities for each observation.}

\item{E_beta, E_beta_sq, E_eta, E_eta_sq}{moment updates produced by \code{\link{get_moments_cpp_doubletree}}}

\item{v1_lookup_NA_replaced}{pL1+1 for \code{NA} in tree1 leaf}

\item{v2_lookup}{\code{v2_lookup} is a vector of length equal to the total number of rows in \code{X};
each element is an integer, indicating which leaf does the observation belong to in tree2}
}
\value{
n by pL1 by K
}
\description{
NB: does this work for the case where missing tree1 label can be in multiple domains,
and missing/non-missing can happen at the same time for any domain?
}
\details{
This function updates the N by pL1 matrix \code{emat} in the package
}
