% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_att_averaged.R
\name{filter_trees_by_structure}
\alias{filter_trees_by_structure}
\title{Filter Trees by Structure}
\usage{
filter_trees_by_structure(tree_list, structure_list, target_structure)
}
\arguments{
\item{tree_list}{List of trees (nested lists for averaging)}

\item{structure_list}{List of TreeStructure objects corresponding to tree_list}

\item{target_structure}{TreeStructure object to match}
}
\value{
List with:
  \item{matched_trees}{List of trees matching target structure}
  \item{n_matched}{Number of matched trees}
}
\description{
Given lists of trees and their structures, return only trees that match
the target structure.
}
\keyword{internal}
