% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_nlcm_doubletree}
\alias{fit_nlcm_doubletree}
\title{Fit nested latent class model with double-tree-structured shrinkage over
\enumerate{
\item domains and 2) causes
}}
\usage{
fit_nlcm_doubletree(
  dsgn,
  vi_params_init,
  hyperparams_init,
  random_init,
  random_init_vals,
  tol,
  tol_hyper,
  max_iter,
  print_freq,
  quiet,
  plot_fig,
  update_hyper_freq,
  hyper_fixed
)
}
\arguments{
\item{dsgn}{a list; data organized according to the trees
(see \code{\link[=design_doubletree]{design_doubletree()}})}

\item{vi_params_init, hyperparams_init, random_init, random_init_vals, tol, tol_hyper, max_iter, print_freq, quiet, plot_fig, update_hyper_freq, hyper_fixed}{initial values and updating protocols. Explained more in the wrapper function
\code{\link[=nlcm_doubletree]{nlcm_doubletree()}}}
}
\value{
a list with model updates. Because the variational posterior is
comprised of familiar distributional forms that can be determined by the moments,
the returned values are these moments:
\describe{
\item{\code{vi_params}}{named list of final variational parameter estimates}
\item{\code{hyperparams}}{named list of final hyperparameter estimates}
\item{\code{hyper_fixed}}{named list of fixed hyperparameters}
\item{\code{ELBO_track}}{numeric vector containing the values of the objective function
(ELBO) at the end of every iteration.}
}
}
\description{
Fit nested latent class model with double-tree-structured shrinkage over
\enumerate{
\item domains and 2) causes
}
}
\concept{internal VI functions}
