% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_line1_2_15_doubletree}
\alias{get_line1_2_15_doubletree}
\title{calculate line 1, 2, and 15 of ELBO*}
\usage{
get_line1_2_15_doubletree(
  F,
  digamma_emat,
  rmat,
  emat,
  v1_lookup_NA_replaced,
  v2_lookup
)
}
\arguments{
\item{F}{intermediate values from updates of rmat and emat; dimension: N by pL1 by K;
currently for observations with known leaf labels in tree1 AND tree2, there is redundancy
because once we know i, we know the leaf 1 label, hence no need for
the second dimension}

\item{digamma_emat}{matrix(pL1 by pL2)}

\item{rmat}{matrix (N by K); row sums are one - the class probabilities for each observation.}

\item{emat}{matrix (N by pL1)}

\item{v1_lookup_NA_replaced}{vector; includes values of pL+1 for unknown leaf label in tree1.}

\item{v2_lookup}{\code{v2_lookup} is a vector of length equal to the total number of rows in \code{X};
each element is an integer, indicating which leaf does the observation belong to in tree2}
}
\value{
line 1, 2, 15 of the ELBO* in Appendix
}
\description{
Intended to be faster than its counterpart implementation in \code{R}.
}
\details{
This function updates the N by pL1 matrix \code{emat} in the package
}
