% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_moments_cpp_doubletree}
\alias{get_moments_cpp_doubletree}
\title{Calculate variational moments during the updates}
\usage{
get_moments_cpp_doubletree(
  prob1,
  prob2,
  mu_gamma,
  sigma_gamma,
  mu_alpha,
  sigma_alpha,
  anc1,
  anc2,
  cardanc1,
  cardanc2
)
}
\arguments{
\item{prob1, prob2}{variational probabilities; \code{prob1} is for \code{s*_u} - length \code{p1};
\code{prob2} is for \code{s_cu} - a matrix \code{pL1} by \code{p2}; in R, a list of pL1 length - each element being of length \code{p2}.}

\item{mu_gamma}{variational Gaussian means (for \code{s*_u=1} component) for J*K
logit(class-specific response probabilities); (J,K,p1) array; In R, we used a list of p1 (J,K) matrices}

\item{sigma_gamma}{variational Gaussian variances (for \code{s*_u=1} component)
for J*K logit(class-specific response probabilities); (J,K,p1) array; in R, we used a list o f p1 (J,K) matrices}

\item{mu_alpha}{variational Gaussian mean vectors (for \code{s_cu=1} component) -
this is a pL1 by K-1 by p2 array; in R, we used a list of p2 matrices (each of dimension pL1 by K-1)}

\item{sigma_alpha}{variational Gaussian variances (for \code{s_cu=1} component)
\itemize{
\item this is an array of dimension (pL1, K-1, p2); in R, we used a list of p2 matrices,
each of dimension pL1 by K-1.
}}

\item{anc1, anc2}{\code{anc1} is a list of pL1 vectors, each vector has the node ids of the ancestors in tree1;
lengths may differ. The ancestors include the node concerned; similarly for \code{anc2}}

\item{cardanc1, cardanc2}{\code{cardanc1} is a numeric vector of length pL1; integers. The number
of ancestors for each leaf node in tree1; similarly for \code{cardanc2}.}
}
\value{
a List

\describe{
return List::create(Named("E_beta")=E_beta,
Named("E_beta_sq")=E_beta_sq,
Named("E_eta")=E_eta,
Named("E_eta_sq")=E_eta_sq);
}
}
\description{
Get all moments that need updating when iterating over internal and leaf nodes;
for both trees. When updating a node u, all the moments
of the descedant nodes will be changed. A recalculation of the moments are necessary
when moving on to another node.
}
