% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_moments_cpp_eco_alpha_doubletree}
\alias{get_moments_cpp_eco_alpha_doubletree}
\title{Calculate variational moments during the updates}
\usage{
get_moments_cpp_eco_alpha_doubletree(
  prob2,
  mu_alpha,
  sigma_alpha,
  E_eta,
  E_eta_sq,
  anc2,
  leaf_desc
)
}
\arguments{
\item{prob2}{\code{prob2} is for \code{s_cu} - a matrix \code{pL1} by \code{p2}; in R, a list of pL1 length - each element being of length \code{p2}.}

\item{mu_alpha}{variational Gaussian mean vectors (for \code{s_cu=1} component) -
this is a pL1 by K-1 by p2 array; in R, we used a list of p2 matrices (each of dimension pL1 by K-1)}

\item{sigma_alpha}{variational Gaussian variances (for \code{s_cu=1} component)
\itemize{
\item this is an array of dimension (pL1, K-1, p2); in R, we used a list of p2 matrices,
each of dimension pL1 by K-1.
}}

\item{E_eta, E_eta_sq}{the intermediate moments to be updated.}

\item{anc2}{\code{anc2} is a list of pL1 vectors, each vector has the node ids of the ancestors in tree2;}

\item{leaf_desc}{leaf descendants under consideration}
}
\value{
a List

\describe{
return List::create(Named("E_eta")=E_eta,
Named("E_eta_sq")=E_eta_sq);
}
}
\description{
Get all moments that need updating when iterating over internal and leaf nodes;
for both trees. When updating a node u, all the moments
of the descendant nodes will be changed. A recalculation of the moments are necessary
when moving on to another node.
}
