% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_moments_cpp_eco_gamma_doubletree}
\alias{get_moments_cpp_eco_gamma_doubletree}
\title{Calculate variational moments during the updates}
\usage{
get_moments_cpp_eco_gamma_doubletree(
  prob1,
  mu_gamma,
  sigma_gamma,
  E_beta,
  E_beta_sq,
  anc1,
  leaf_desc
)
}
\arguments{
\item{prob1}{variational probabilities; \code{prob1} is for \code{s*_u} - length \code{p1}.}

\item{mu_gamma}{variational Gaussian means (for \code{s*_u=1} component) for J*K
logit(class-specific response probabilities); (J,K,p1) array; In R, we used a list of p1 (J,K) matrices}

\item{sigma_gamma}{variational Gaussian variances (for \code{s*_u=1} component)
for J*K logit(class-specific response probabilities); (J,K,p1) array; in R, we used a list o f p1 (J,K) matrices}

\item{E_beta, E_beta_sq}{J by K by pL1
each of dimension pL1 by K-1.}

\item{anc1}{\code{anc1} is a list of pL1 vectors, each vector has the node ids of the ancestors in tree1;
lengths may differ. The ancestors include the node concerned; similarly for \code{anc2}}

\item{leaf_desc}{an integer vector of leaf descendant ids under consideration}
}
\value{
a List

\describe{
return List::create(Named("E_beta")=E_beta,
Named("E_beta_sq")=E_beta_sq);
}
}
\description{
Get all moments that need updating when iterating over internal and leaf nodes;
for both trees. When updating a node u, all the moments
of the descendant nodes will be changed. A recalculation of the moments are necessary
when moving on to another node.
}
