% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_topk_COD}
\alias{get_topk_COD}
\title{get top k ids for each row}
\usage{
get_topk_COD(probs, s)
}
\arguments{
\item{probs}{a matrix where columns are causes; the entries are probabilities of causes}

\item{s}{top s}
}
\value{
data frame; for each row in probs, indicating the top s causes
}
\description{
get top k ids for each row
}
\examples{

xx <- matrix(c(0.3,0.1,0.2,0.4,
               0.4,0.1,0.3,0.2),nrow=2,ncol=4,byrow=TRUE)
get_topk_COD(xx,2)

}
