% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCSMFacc}
\alias{getCSMFacc}
\title{get CSMF accuracy}
\usage{
getCSMFacc(csmf, truth, undet = NULL)
}
\arguments{
\item{csmf}{a vector that sums to one; positive values}

\item{truth}{a vector that sums to one; positive values}

\item{undet}{\code{NULL} default.}
}
\value{
numeric; between 0 and 1; the higher the more accurate.
}
\description{
get CSMF accuracy
}
