% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myrdirich}
\alias{myrdirich}
\title{generate random vectors from Dirichlet distribution}
\usage{
myrdirich(n, v)
}
\arguments{
\item{n}{sample size}

\item{v}{a vector of positive values}
}
\value{
matrix of n rows and \code{length(v)} columns; each row sums to 1
}
\description{
generate random vectors from Dirichlet distribution
}
\examples{

myrdirich(10,c(1,1,1))

}
