% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_averaging.R
\name{predict_single_obs}
\alias{predict_single_obs}
\title{Predict for Single Observation}
\usage{
predict_single_obs(tree_node, x_row)
}
\arguments{
\item{tree_node}{Current node (nested list)}

\item{x_row}{Single-row data.frame}
}
\value{
Numeric P(Y=1)
}
\description{
Traverse tree to get prediction for one observation.
}
\keyword{internal}
