% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_att_msplit.R
\name{print.msplit_att}
\alias{print.msplit_att}
\title{Print Method for msplit_att}
\usage{
\method{print}{msplit_att}(x, ...)
}
\arguments{
\item{x}{msplit_att object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for msplit_att
}
