% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{print.summary.nlcm_doubletree_compact}
\alias{print.summary.nlcm_doubletree_compact}
\title{Compact printing of \code{\link[=nlcm_doubletree]{nlcm_doubletree()}} model fits}
\usage{
\method{print}{summary.nlcm_doubletree_compact}(
  x,
  print_leaves = TRUE,
  print_pi = FALSE,
  digits = max(3L, getOption("digits") - 3L),
  ...
)
}
\arguments{
\item{x}{output from \code{summary.nlcm_doubletree} with \code{compact = TRUE}.}

\item{print_leaves}{If \code{TRUE}, for each discovered group the full list
of leaves will be printed. Set this to \code{FALSE} if these leaf lists
make output difficult to read.}

\item{print_pi}{If \code{TRUE} print inference about pi; default to \code{FALSE}.}

\item{digits}{Number of significant digits to print.}

\item{...}{Not used.}
}
\value{
see \code{\link[=print.nlcm_doubletree]{print.nlcm_doubletree()}}
}
\description{
\code{print.summary.nlcm_doubletree_compact} is a print method for class
\code{summary.nlcm_doubletree_compact}.
}
\seealso{
Other nlcm_doubletree results: 
\code{\link{print.summary.nlcm_doubletree_long}()},
\code{\link{summary.nlcm_doubletree}()}
}
\concept{nlcm_doubletree results}
