% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{print.summary.nlcm_doubletree_long}
\alias{print.summary.nlcm_doubletree_long}
\title{Long-form printing of \code{\link[=nlcm_doubletree]{nlcm_doubletree()}} model fits}
\usage{
\method{print}{summary.nlcm_doubletree_long}(x, print_leaves = TRUE, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{output from \code{summary.nlcm_doubletree} with \code{compact = FALSE}.}

\item{print_leaves}{If \code{TRUE}, for each discovered group the full list
of leaves will be printed. Set this to \code{FALSE} if these leaf lists
make output difficult to read.}

\item{digits}{Number of significant digits to print.}

\item{...}{Not used.}
}
\value{
see \code{\link[=print.nlcm_doubletree]{print.nlcm_doubletree()}}
}
\description{
\code{print.summary.nlcm_doubletree_long} is a print method for class
\code{summary.nlcm_doubletree_long}.
}
\seealso{
Other nlcm_doubletree results: 
\code{\link{print.summary.nlcm_doubletree_compact}()},
\code{\link{summary.nlcm_doubletree}()}
}
\concept{nlcm_doubletree results}
