% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_msplit.R
\name{summary.msplit_att}
\alias{summary.msplit_att}
\title{Summary Method for msplit_att}
\usage{
\method{summary}{msplit_att}(object, ...)
}
\arguments{
\item{object}{msplit_att object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for msplit_att
}
