% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{summary.nlcm_doubletree}
\alias{summary.nlcm_doubletree}
\title{\code{summary.nlcm_doubletree} summarizes the results from \code{\link[=nlcm_doubletree]{nlcm_doubletree()}}.}
\usage{
\method{summary}{nlcm_doubletree}(object, compact = FALSE, B = 10000, ...)
}
\arguments{
\item{object}{Output from \code{\link[=nlcm_doubletree]{nlcm_doubletree()}}; of class \code{nlcm_doubletree}
An object of class "lcm_tree".# coeff_type Either "lcm_tree" or "ad_hoc"}

\item{compact}{If \code{TRUE}, a more compact summary of results is printed.
Only works well when the dimension of the variables is low
(say, < 4); otherwise the table takes up too much horizontal space.
Default is \code{FALSE}.}

\item{B}{Default \code{10000}: The number of random multivariate Gaussian for logit of V in stick breaking
parameterization; used for obtaining the posterior distribution of \verb{lambda^(c,g)_k}}

\item{...}{Not used.}
}
\value{
see \code{\link[=print.nlcm_doubletree]{print.nlcm_doubletree()}}
}
\description{
Will have some tiny randomness associated with the upper and lower bounds,
because we simulated the Gaussian variables before converting to probabilities.
}
\seealso{
Other nlcm_doubletree results: 
\code{\link{print.summary.nlcm_doubletree_compact}()},
\code{\link{print.summary.nlcm_doubletree_long}()}
}
\concept{nlcm_doubletree results}
