% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_alpha_subid_doubletree}
\alias{update_alpha_subid_doubletree}
\title{Update alpha's variational distribution.}
\usage{
update_alpha_subid_doubletree(
  u,
  v1,
  g_psi,
  g_phi,
  tau_2_t_u,
  E_eta,
  E_xi_u,
  X,
  rmat,
  emat,
  h_pau,
  levels,
  subject_ids,
  v2_lookup
)
}
\arguments{
\item{u}{node id; internal or leaf node in tree1}

\item{v1}{leaf node id in tree1.}

\item{g_psi, g_phi}{g of local variational parameters}

\item{tau_2_t_u}{variational Gaussian variances for gamma}

\item{E_eta, E_xi_u}{moment updates produced by \code{\link{get_moments_cpp}};
\code{E_xi_u} is directly calculated}

\item{X}{transformed data: \verb{2Y-1}; contains potential missing data.}

\item{rmat}{a matrix of variational probabilities of all observations
belong to K classes; N by K; each row sums to 1}

\item{emat}{a matrix of variational probability for all observations
belonging to pL1 leaf nodes; N by pL1; each row sums to 1. Importantly,
for rows with obsered leaf nodes in tree1, we just have an one-hot represention
of that cause.}

\item{h_pau}{a numeric vector of length p indicating the branch length
between a node and its parent; for tree2}

\item{levels}{a vector of possibly repeating integers from 1 to Fg2}

\item{subject_ids}{integer ids for subjects nested under node \code{u}}

\item{v2_lookup}{a vector of length equal to the total number of rows in X;
each element is an integer, indicating which leaf does the observation belong to in tree2.}
}
\value{
a list
\describe{
\item{resC}{actually 1/C in the paper, this is variance}
\item{resD}{}
\item{logresDsq_o_C}{}
}
}
\description{
Update alpha's variational distribution.
}
