% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_emat_with_F_doubletree}
\alias{update_emat_with_F_doubletree}
\title{Update the variational probabilities of each observation in one of \code{pL1} leaves in tree1}
\usage{
update_emat_with_F_doubletree(curr_F, rmat, digamma_emat, v2_lookup)
}
\arguments{
\item{curr_F}{F from \code{\link[=F_doubletree]{F_doubletree()}}}

\item{rmat}{matrix (N by K); row sums are one - the class probabilities for each observation.}

\item{digamma_emat}{matrix(pL1 by pL2)}

\item{v2_lookup}{\code{v2_lookup} is a vector of length equal to the total number of rows in \code{X};
each element is an integer, indicating which leaf does the observation belong to in tree2}
}
\value{
N by pL1 variational multinomial probabilities for
the posterior distribution of N people's membership in the pL1 leaves in tree1;
row sums are 1s. \code{F} is to prepare for \code{rmat} updates. For subjects with
observed leaf label in tree1, the corresponding rows of \code{res} should be further modified.
}
\description{
NB: does this work for the case where missing tree1 label can be in multiple domains,
and missing/non-missing can happen at the same time for any domain?
}
\details{
This function updates the N by pL1 matrix \code{emat} in the package
}
