% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_gamma_subid_doubletree}
\alias{update_gamma_subid_doubletree}
\title{Update gamma's variational distribution. Update the variational mean and variance for logit of
class-specific response probabilities (for the \code{s*_u=1} component in tree1.)}
\usage{
update_gamma_subid_doubletree(
  u,
  g_psi,
  g_phi,
  tau_1_t_u,
  E_beta,
  E_zeta_u,
  X_zeropad,
  rmat,
  emat,
  h_pau,
  subject_ids_nonmissing,
  leaf_desc
)
}
\arguments{
\item{u}{node id; internal or leaf node in tree1}

\item{g_psi, g_phi}{g of local variational parameters}

\item{tau_1_t_u}{variational Gaussian variances for gamma}

\item{E_beta, E_zeta_u}{moment updates produced by \code{\link{get_moments_cpp}};
\code{E_zeta_u} is directly calculated: \code{prob1[u]*sigma_gamma[u,,]}}

\item{X_zeropad}{transformed data: \verb{2Y-1}; contains potential missing data.}

\item{rmat}{a matrix of variational probabilities of all observations
belong to K classes; N by K; each row sums to 1}

\item{emat}{a matrix of variational probability for all observations
belonging to pL1 leaf nodes; N by pL1; each row sums to 1. Importantly,
for rows with obsered leaf nodes in tree1, we just have an one-hot represention
of that cause.}

\item{h_pau}{a numeric vector of length p indicating the branch length
between a node and its parent}

\item{subject_ids_nonmissing}{the ids of subjects in the leaf descendants of node u;
a list of length J, each is a list of subjects nested under u AND have complete info.}

\item{leaf_desc}{a vector of leaf descendants nested under node \code{u}}
}
\value{
a list
\describe{
\item{resA}{actually 1/A in the paper, this is variance}
\item{resB}{}
\item{logresBsq_o_A}{}
}
}
\description{
Update gamma's variational distribution. Update the variational mean and variance for logit of
class-specific response probabilities (for the \code{s*_u=1} component in tree1.)
}
