% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_hyperparams_doubletree}
\alias{update_hyperparams_doubletree}
\title{update hyperparameters}
\usage{
update_hyperparams_doubletree(
  h_pau,
  levels,
  v_units,
  X,
  n,
  J,
  p1,
  p2,
  pL1,
  pL2,
  Fg1,
  Fg2,
  ind_obs_i,
  prob1,
  prob2,
  mu_gamma,
  mu_alpha,
  rmat,
  emat,
  dirich_mat,
  sigma_gamma,
  sigma_alpha,
  tau_1_t,
  tau_2_t,
  a1_t,
  b1_t,
  a2_t,
  b2_t,
  E_beta_sq,
  E_eta_sq,
  E_beta,
  E_eta,
  psi,
  g_psi,
  phi,
  g_phi,
  tau_1,
  tau_2,
  a1,
  b1,
  a2,
  b2,
  dmat,
  K,
  LD,
  tau_update_levels,
  update_hyper,
  quiet,
  do_tree1_update
)
}
\arguments{
\item{prob1, prob2, mu_gamma, mu_alpha, rmat, emat, dirich_mat, sigma_gamma, sigma_alpha, tau_1_t, tau_2_t, a1_t, b1_t, a2_t, b2_t}{variational parameters}

\item{E_beta_sq, E_eta_sq, E_beta, E_eta}{moments computed by \code{\link[=update_vi_params_doubletree]{update_vi_params_doubletree()}}}

\item{psi, g_psi, phi, g_phi, tau_1, tau_2}{parameters updated by \code{update_hyperparams_doubletree()}}

\item{tau_update_levels}{a numeric vector, specifies which levels of hyperparameters to update}

\item{update_hyper}{Logical, \code{TRUE} or \code{FALSE} to indicate
whether to update \code{tau_1} and \code{tau_2}. This is computed at every iteration
in \code{\link[=fit_nlcm_doubletree]{fit_nlcm_doubletree()}}}

\item{quiet}{default to \code{FALSE}, which prints intermediate updates of hyperparameters}

\item{do_tree1_update}{\code{TRUE} to update \code{mu_gamma},\code{sigma_gamma}, \code{prob1},\code{tau_1}}
}
\value{
a list of updated hyperparameters: tau_1,tau_2,psi,g_psi,phi,g_phi,
along with a new ELBO value.
}
\description{
NB: argument redundancy may exist
}
\seealso{
Other Internal VI functions: 
\code{\link{update_vi_params_doubletree}()}
}
\concept{Internal VI functions}
