% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_rmat_with_F_doubletree}
\alias{update_rmat_with_F_doubletree}
\title{Update the variational probabilities of each observation in one of \code{K} classes}
\usage{
update_rmat_with_F_doubletree(curr_F, emat)
}
\arguments{
\item{curr_F}{from \code{\link{F_doubletree}}}

\item{emat}{matrix (N by K); row sums are one - the class probabilities for each observation.}
}
\value{
N by K variational multinomial probabilities for
the posterior distribution of N people's membership in the K classes in tree1;
row sums are 1s.
}
\description{
This function updates the N by K matrix \code{rmat} in the package
}
