% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_vi_params_doubletree}
\alias{update_vi_params_doubletree}
\title{update variational parameters}
\usage{
update_vi_params_doubletree(
  leaf_ids_units,
  leaf_ids_nodes,
  ancestors,
  h_pau,
  levels,
  v_units,
  subject_id_list,
  ind_obs_i,
  ind_obs_j,
  scenario,
  X,
  n,
  J,
  p1,
  p2,
  pL1,
  pL2,
  Fg1,
  Fg2,
  cardanc1,
  cardanc2,
  prob1,
  prob2,
  mu_gamma,
  mu_alpha,
  rmat,
  emat,
  dirich_mat,
  sigma_gamma,
  sigma_alpha,
  tau_1_t,
  tau_2_t,
  a1_t,
  b1_t,
  a2_t,
  b2_t,
  psi,
  g_psi,
  phi,
  g_phi,
  tau_1,
  tau_2,
  a1,
  b1,
  a2,
  b2,
  dmat,
  K,
  LD,
  s1_u_zeroset,
  s1_u_oneset,
  s2_cu_zeroset,
  s2_cu_oneset
)
}
\arguments{
\item{leaf_ids_units, leaf_ids_nodes, ancestors, h_pau, levels, v_units, subject_id_list, ind_obs_i, ind_obs_j, scenario}{outputs from \code{\link[=design_doubletree]{design_doubletree()}}. Among them,
\code{subject_id_list} is currently not used in the initialization
function \code{\link[=initialize_nlcm_doubletree]{initialize_nlcm_doubletree()}}, but are useful for updating functions here.}

\item{X, n, J, p1, p2, pL1, pL2, Fg1, Fg2, cardanc1, cardanc2}{data and design information computed from the outputs
of \code{\link[=initialize_nlcm_doubletree]{initialize_nlcm_doubletree()}}; we use \code{R.utils::do.Call} to get their values from \code{dsgn}.}

\item{prob1, prob2, mu_gamma, mu_alpha, rmat, emat, dirich_mat, sigma_gamma, sigma_alpha, tau_1_t, tau_2_t, a1_t, b1_t, a2_t, b2_t}{variational parameters}

\item{psi, g_psi, phi, g_phi, tau_1, tau_2}{parameters updated by \code{update_hyperparams_doubletree()}}

\item{a1, b1, a2, b2, dmat, K, LD, s1_u_zeroset, s1_u_oneset, s2_cu_zeroset, s2_cu_oneset}{fixed hyperparameters}
}
\description{
Used by \code{\link[=fit_nlcm_doubletree]{fit_nlcm_doubletree()}}; followed by \code{\link[=update_hyperparams_doubletree]{update_hyperparams_doubletree()}}
}
\seealso{
Other Internal VI functions: 
\code{\link{update_hyperparams_doubletree}()}
}
\concept{Internal VI functions}
