% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletree_example_data.R
\docType{data}
\name{X0}
\alias{X0}
\title{example VA data from PHMRC - adult}
\format{
matrix; the length is 7841 by 168
}
\usage{
X0
}
\description{
a matrix of data with potential missing data
\code{1}: yes; \code{0}: no; \code{NA}: missing.
The rows of this data match with \code{cause_ids}
and \code{domain_ids}
}
\keyword{datasets}
